Remove opta for rugby
